// resources/js/types/product.ts
// Generated by https://quicktype.io

export interface Product {
    id: number;
    kategori_id: number;
    satuan_id: number;
    nama_produk: string;
    slug: string;
    deskripsi: string;
    harga_beli: number;
    harga_jual: number;
    promo: number;
    percentage: number;
    stok: number;
    is_active: number;
    image: null | string;
    created_at: Date;
    updated_at: Date;
    kategori: Kategori;
    satuan: Satuan;
}

export interface Kategori {
    id: number;
    kategori: string;
    slug: string;
    icon: string;
    created_at: Date;
    updated_at: Date;
}

export interface Satuan {
    id: number;
    satuan: string;
    created_at: Date;
    updated_at: Date;
}