// resources/js/types/setting.ts
// Generated by https://quicktype.io

export interface Setting {
    id: number;
    site_name: string;
    keyword: string;
    description: string;
    email: string;
    phone: string;
    address: string;
    fb: string;
    ig: string;
    tiktok: string;
    ppn: number;
    biaya_lainnya: number;
    logo: string;
    created_at: null;
    updated_at: Date;
}